Move Context class from Unify to its own file
The Context class (nothing to do with ContextLink) hold the quotation
and the shadowing variables of atoms.
