+++ opencog/atoms/base/CMakeLists.txt
@@ -20,6 +20,7 @@ ADD_LIBRARY (atombase
 	LinkValue.cc
 	Node.cc
 	Quotation.cc
+	Context.cc
 	StringValue.cc
 	Valuation.cc
 )
@@ -48,6 +49,7 @@ INSTALL (FILES
 	Node.h
 	ProtoAtom.h
 	Quotation.h
+	Context.h
 	StringValue.h
 	types.h
 	Valuation.h
